feat: move block and interfaces to encoding
