--- packages/core/src/network/index.ts
@@ -1,3 +1,0 @@
-export * from "./event"
-export * from "./network"
-export * from "./record"
