+++ packages/encoding/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./coding"
 export * from "./event"
+export * from "./interfaces"
+export * from "./ipld"
 export * from "./record"
