+++ packages/encoding/src/interfaces.ts
@@ -1,8 +1,53 @@
-import { ThreadID } from "@textile/threads-id"
-import CID from "cids"
-import { Block } from "../ipld"
-import { LogID } from "../thread"
-import { Event } from "./event"
+import type { LogID } from "@textile/threads-core"
+import type { ThreadID } from "@textile/threads-id"
+import type CID from "cids"
+import type { Block } from "./ipld"
+
+/**
+ * Event is a Block node representing an event.
+ */
+export interface Event {
+  /**
+   * The node structure of the event.
+   */
+  value: Block<EventNode>
+  /**
+   * The header content for the event.
+   */
+  header: Block<Uint8Array>
+  /**
+   * The body content for the event.
+   */
+  body: Block<Uint8Array>
+  /**
+   * The underlying event node.
+   */
+  obj?: EventNode
+}
+
+/**
+ * Node defines the node structure of an event.
+ */
+export interface EventNode {
+  /**
+   * CID of body block
+   */
+  body: CID
+  /**
+   * CID of header block
+   */
+  header: CID
+}
+
+/**
+ * Header defines the node structure of an event header.
+ */
+export interface EventHeader {
+  /**
+   * Single-use symmetric key
+   */
+  key?: Uint8Array
+}
 
 /**
  * LogRecord is an Block node representing a record.
